Fix XMLRPCServer setup and callback registration for seperate internal and external IPs.
When pyhomematic is run in a docker container a common network setup is that the external
reachable IP is different from the local container IP. In this case the XMLRPCServer could not start.

Instead binding to "local" we bind to 0.0.0.0 (all interfaces).
RPC callback registration uses now "local" as callback address.