Added Support for HmIP-SWO-PL
I manually edited the sensors.py on my system to get my "HmIP-SWO-PL" working in Home Assistant. 

It is nearly the same as the "HmIP-SWO-PR": IPWeatherSensor which is already supported with pyhomematic just without the wind direction sensor.  
https://www.eq-3.de/produkte/homematic-ip/homematic-ip-wettersensor-plus.html
The two values for "WIND_DIR" and "WIND_DIR_RANGE" now show "false" as value. Maybe there is a smart way to not create those sensors if they are not available?
Thx