New master classes and Helper usage
I have changed the use of main classes and it inheritance. The new top
class SensorHm, SensorHmIP include all Helper classes that are common
to all devices of the specific model line. SensorHm is for standard
Homematic devices and SensorHmIP for the newer IP version.

Also, the use of existing Helper classes is now implemented for all
sensor classes. to prepare for the nex step of adding more Helper
classes and changing the class names I have partly copied the content of
a sensor class to a class that was until now inheriting a sensor class
of similar capability. An example is the Shutter class.